C99 types replacement + 64bits support.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@1247 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ libhxcfe/sources/fs_manager/fs_cpm/fs_cpm.h
@@ -13,7 +13,7 @@ int cpm_createDir( HXCFE_FSMNG * fsmng,char * foldername);
 int cpm_removeDir( HXCFE_FSMNG * fsmng,char * foldername);
 
 int cpm_ftell( HXCFE_FSMNG * fsmng,int filehandle);
-int cpm_fseek( HXCFE_FSMNG * fsmng,int filehandle,long offset,int origin);
+int cpm_fseek( HXCFE_FSMNG * fsmng,int filehandle,int32_t offset,int origin);
 
 int cpm_openDir(HXCFE_FSMNG * fsmng, char * path);
 int cpm_readDir(HXCFE_FSMNG * fsmng,int dirhandle,HXCFE_FSENTRY * dirent);
