+++ libhxcfe/sources/loaders/afi_loader/afi_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int AFI_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int AFI_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
