+++ libhxcfe/sources/loaders/emui_raw_loader/emui_raw_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int EMUI_RAW_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int EMUI_RAW_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
