+++ libhxcfe/sources/loaders/krz_loader/krz_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int KRZ_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int KRZ_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
