+++ libhxcfe/sources/loaders/sdd_speccydos_loader/sdd_speccydos_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int SDDSpeccyDos_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int SDDSpeccyDos_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
