+++ libhxcfe/sources/loaders/stt_loader/stt_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int STT_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int STT_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
