+++ libhxcfe/sources/thirdpartylibs/FATIOlib/example.c
@@ -7,9 +7,9 @@ int media_init()
     return 1;
 }
 
-int media_read(unsigned long sector, unsigned char *buffer, unsigned long sector_count)
+int media_read(uint32_t sector, unsigned char *buffer, uint32_t sector_count)
 {
-    unsigned long i;
+    uint32_t i;
 
     for (i=0;i<sector_count;i++)
     {
@@ -24,9 +24,9 @@ int media_read(unsigned long sector, unsigned char *buffer, unsigned long sector
     return 1;
 }
 
-int media_write(unsigned long sector, unsigned char *buffer, unsigned long sector_count)
+int media_write(uint32_t sector, unsigned char *buffer, uint32_t sector_count)
 {
-    unsigned long i;
+    uint32_t i;
 
     for (i=0;i<sector_count;i++)
     {
